prototyping
This version uses deck.gl + Leaflet with a raster basemap, as described
at <https://deck.gl/docs/get-started/using-with-map>.

Pro/con: Rendering is paused when zooming, panning, etc. On the one hand
this ensures that interactivity never suffers, but it does look little
weird.

Using a raster basemap is desirable so as to use the familiar osm-carto
style that Overpass Turbo users are used to. It also may be faster but
this is pure speculation.

The Stamen Toner tiles are extremely small in raster format and load
very quickly (100x smaller than the equivalent vector tiles). The black
and white style is useful for visualizing data on top of. So it would be
nice to offer both.

Will probably end up using Maplibre GL JS with a raster basemap, but we
can easily switch to Leaflet as an option if there's a good reason to.