prototyping
Switch from Leaflet to MapLibre, again using the Deck examples.

Interactivity is significantly improved in that both layers move
together, but ironically this suffers on extremely large datasets as it
redraws every frame instead of panning around a single lightweight
raster.

Interactivity in this example is controlled by Deck, not by MapLibre, so
for instance double tap and drag is missing. The interleaved layer would
not have this issue, although they describe it as being potentially less
robust.

Their examples include react-map-gl which depends on Mapbox; they give
an example of using Webpack to rewrite the dependencies but I didn't
attempt to work it. So this is not hooked up with React.